Add the ability to register dataporten applications.
A new dataporten application can now be created when installing a new
release. The client ID, client secret and owner is returned to the user.
Currently nothing will happen if the registration succeds, but the
installation of the release fails.
The dataporten application should probably be deleted if the release
could not be properly installed.